Gemini review feedback

+++ packages/video_player/video_player_android/android/src/main/java/io/flutter/plugins/videoplayer/VideoPlayer.java
@@ -117,7 +117,7 @@ public abstract class VideoPlayer implements Messages.VideoPlayerInstanceApi {
 
   @Override
   public void seekTo(@NonNull Long position) {
-    exoPlayer.seekTo(position.intValue());
+    exoPlayer.seekTo(position);
   }
 
   @NonNull
